+++ tests/testthat/test_lulu.R
@@ -102,7 +102,8 @@ testthat::test_that("lulu_map_impl works", {
       0.16,
       min_cooccurrence_ratio = 0.95,
       min_abundance_ratio = 0.9,
-      use_mean_abundance_ratio = TRUE
+      use_mean_abundance_ratio = TRUE,
+      verbose = 0
     ),
     tibble::tibble(
       seq_idx = 1L:12L,
@@ -112,15 +113,82 @@ testthat::test_that("lulu_map_impl works", {
   )
 })
 
+testthat::test_that("lulu_map_lowmem_impl works", {
+  comm_matrix_long$seq_idx <- as.integer(comm_matrix_long$seq_id)
+  pairs$seq_idx1 <- as.integer(pairs$seq_id1)
+  pairs$seq_idx2 <- as.integer(pairs$seq_id2)
+  targets::tar_dir({
+    readr::write_tsv(pairs, "pairs.tsv")
+    readr::write_tsv(comm_matrix_long, "comm_matrix_long.tsv")
+    targets::tar_script({
+      library('targets')
+      library('tarchetypes')
+      list(
+        tar_fst_tbl(
+          comm_matrix_long1,
+          readr::read_tsv('comm_matrix_long.tsv', col_types = "icii") |>
+            dplyr::slice_head(prop = 0.5)
+        ),
+        tar_fst_tbl(
+          comm_matrix_long2,
+          readr::read_tsv('comm_matrix_long.tsv', col_types = "icii") |>
+            dplyr::slice_tail(prop = 0.5)
+        ),
+        tar_fst_tbl(
+          pairs1,
+          readr::read_tsv('pairs.tsv', col_types = "icicidii") |>
+            dplyr::slice_head(prop = 0.5)
+        ),
+        tar_fst_tbl(
+          pairs2,
+          readr::read_tsv('pairs.tsv', col_types = "icicidii") |>
+            dplyr::slice_tail(prop = 0.5)
+        ),
+        tar_fst_tbl(
+          lulu_map,
+          {
+            # verify that the dependencies have *not* been loaded by targets
+            stopifnot(isFALSE(exists("comm_matrix_long1")))
+            stopifnot(isFALSE(exists("comm_matrix_long2")))
+            stopifnot(isFALSE(exists("pairs1")))
+            stopifnot(isFALSE(exists("pairs2")))
+            optimotu.pipeline::lulu_map_lowmem(
+              otu_table = c(comm_matrix_long1, comm_matrix_long2),
+              match_table = c(pairs1, pairs2),
+              max_dist = 0.16,
+              min_cooccurrence_ratio = 0.95,
+              min_abundance_ratio = 0.9,
+              use_mean_abundance_ratio = FALSE,
+              verbose = 0
+            )
+          },
+          retrieval = "none"
+        )
+      )
+    })
+    targets::tar_make(callr_function = NULL, reporter = "silent")
+    testthat::expect_equal(
+      targets::tar_read(lulu_map),
+      tibble::tibble(
+        seq_idx = 1L:12L,
+        lulu_idx = c(1L, 3L, 9L)[as.integer(substring(colnames(comm_matrix), 2L))]
+      ),
+      ignore_attr = TRUE
+    )
+  })
+})
+
 # Run example from LULU github
 
 download.file(
   "https://raw.githubusercontent.com/tobiasgf/lulu/master/Example_data/centroids_test.txt",
-  destfile = "centroids_test.txt"
+  destfile = "centroids_test.txt",
+  quiet = TRUE
 )
 download.file(
   "https://raw.githubusercontent.com/tobiasgf/lulu/master/Example_data/otutable_test.txt",
-  destfile = "otutable_test.txt"
+  destfile = "otutable_test.txt",
+  quiet = TRUE
 )
 
 # run in terminal
@@ -211,6 +279,7 @@ long_lulu_map <- optimotu.pipeline::lulu_map(
   verbose = 0
 )
 
+
 testthat::test_that("lulu_map and lulu give same map result", {
   testthat::expect_equal(
     long_lulu_map$lulu_id,
@@ -258,4 +327,98 @@ testthat::test_that("lulu_table and lulu give same result", {
     wide_lulu_otutab,
     lulu_result$curated_table
   )
+})
+
+testthat::test_that("lulu_map_lowmem gives the same result as lulu_map", {
+  targets::tar_dir({
+    # we need to convert from seq_id to seq_idx
+    seq_ids <- unique(long_otutab$seq_id)
+    dplyr::mutate(
+      long_otutab,
+      seq_idx = as.integer(factor(seq_id, levels = seq_ids)),
+      .keep = "unused"
+    ) |>
+      saveRDS("long_otutab.rds")
+    dplyr::mutate(
+      long_matchlist,
+      seq_idx1 = as.integer(factor(seq_id1, levels = seq_ids)),
+      seq_idx2 = as.integer(factor(seq_id2, levels = seq_ids))
+    ) |>
+      saveRDS("long_matchlist.rds")
+
+    targets::tar_script({
+      library(targets)
+      library(tarchetypes)
+
+      # we use different types of formats and branching for the purposes of
+      # testing
+      list(
+        # first part is just a targets (similar to static branching)
+        tar_target(
+          otu_table1,
+          readRDS("long_otutab.rds") |>
+            dplyr::slice_head(prop = 0.5)
+        ),
+        # second part is dynamically branched
+        tar_group_count(
+          otu_table2,
+          readRDS("long_otutab.rds") |>
+            dplyr::slice_tail(prop = 0.5),
+          count = 3
+        ),
+        # realize the dynamic branches; use a different format
+        tar_target(
+          otu_table2_branch,
+          otu_table2,
+          pattern = map(otu_table2),
+          format = "fst"
+        ),
+        # first part a single target, another format
+        tar_target(
+          match_table1,
+          readRDS("long_matchlist.rds") |>
+            dplyr::slice_head(prop = 0.5),
+          format = "qs"
+        ),
+        # second part branched
+        tar_group_size(
+          match_table2,
+          readRDS("long_matchlist.rds") |>
+            dplyr::slice_tail(prop = 0.5),
+          size = 50000
+        ),
+        # realize the branches in a different format
+        tar_target(
+          match_table2_branch,
+          match_table2,
+          pattern = map(match_table2),
+          format = "parquet"
+        ),
+        tar_target(
+          lulu_map,
+          optimotu.pipeline::lulu_map_lowmem(
+            otu_table = list(otu_table1, otu_table2_branch),
+            # cbind is semantically wrong, but it is ignored here!
+            match_table = cbind(match_table1, match_table2_branch),
+            max_dist = 0.16,
+            min_abundance_ratio = 1,
+            min_cooccurrence_ratio = 1,
+            use_mean_abundance_ratio = FALSE,
+            verbose = 0
+          )
+        )
+      )
+    })
+    targets::tar_make(reporter = "silent", callr_function = NULL)
+    long_lulu_map_idx <- dplyr::transmute(
+      long_lulu_map,
+      seq_idx = as.integer(factor(seq_id, levels = seq_ids)),
+      lulu_idx = as.integer(factor(lulu_id, levels = seq_ids))
+    ) |>
+      dplyr::arrange(seq_idx)
+    testthat::expect_equal(
+      targets::tar_read(lulu_map) |> dplyr::arrange(seq_idx),
+      long_lulu_map_idx
+    )
+  })
 })